3.2 Adding Grabbable Game Objects
Pretty Straightforward. Most important thing to keep track of is the use of XRI Default Left and Right Controller presets for the direct interactors and to disable the position, rotation, is tracked, and tracking state actions for the left and right tp controllers when making them a child of the direct interactors.